requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0

